refactored S3Storage: S3 operations
