+++ save-cloud-common/src/jvmMain/kotlin/com/saveourtool/save/storage/AbstractS3Storage.kt
@@ -1,18 +1,15 @@
 package com.saveourtool.save.storage
 
+import com.saveourtool.save.s3.S3Operations
 import com.saveourtool.save.utils.debug
 import com.saveourtool.save.utils.getLogger
 
 import org.slf4j.Logger
-import org.springframework.http.MediaType
 import org.springframework.web.reactive.function.BodyExtractors.toFlux
 import reactor.core.publisher.Flux
 import reactor.core.publisher.Mono
 import reactor.kotlin.core.publisher.toFlux
 import reactor.kotlin.core.publisher.toMono
-import reactor.kotlin.core.util.function.component1
-import reactor.kotlin.core.util.function.component2
-import software.amazon.awssdk.core.async.AsyncRequestBody
 import software.amazon.awssdk.core.async.AsyncResponseTransformer
 import software.amazon.awssdk.services.s3.S3AsyncClient
 import software.amazon.awssdk.services.s3.model.*
@@ -23,156 +20,59 @@ import java.time.Instant
 /**
  * S3 implementation of Storage
  *
- * @param s3Client async S3 client to operate with S3 storage
- * @param bucketName
- * @param prefix a common prefix for all keys in S3 storage for this storage
+ * @param s3Operations [S3Operations] to operate with S3
+ * @param prefix a common prefix for all S3 keys in this storage
  * @param K type of key
  */
 abstract class AbstractS3Storage<K>(
     private val s3Client: S3AsyncClient,
-    private val bucketName: String,
+    private val s3Operations: S3Operations,
     prefix: String,
 ) : Storage<K> {
     private val log: Logger = getLogger(this::class)
     private val prefix = prefix.removeSuffix(PATH_DELIMITER) + PATH_DELIMITER
 
-    override fun list(): Flux<K> = listObjectsV2()
+    override fun list(): Flux<K> = s3Operations.listObjectsV2(prefix)
         .flatMapIterable { response ->
             response.contents().map {
                 buildKey(it.key().removePrefix(prefix))
             }
         }
 
-    private fun listObjectsV2(): Flux<ListObjectsV2Response> = doListObjectsV2().expand { lastResponse ->
-        if (lastResponse.isTruncated) {
-            doListObjectsV2(lastResponse.nextContinuationToken())
-        } else {
-            Mono.empty()
-        }
-    }
-
-    private fun doListObjectsV2(continuationToken: String? = null): Mono<ListObjectsV2Response> = ListObjectsV2Request.builder()
-        .bucket(bucketName)
-        .prefix(prefix)
-        .let { builder ->
-            continuationToken?.let { builder.continuationToken(it) } ?: builder
-        }
-        .build()
-        .let {
-            s3Client.listObjectsV2(it).toMono()
+    override fun download(key: K): Flux<ByteBuffer> = s3Operations.getObject(buildS3Key(key))
+        .flatMapMany {
+            it.toFlux()
         }
 
-    override fun download(key: K): Flux<ByteBuffer> {
-        val request = GetObjectRequest.builder()
-            .bucket(bucketName)
-            .key(buildS3Key(key))
-            .build()
+    override fun upload(key: K, content: Flux<ByteBuffer>): Mono<Long> =
+            s3Operations.uploadObject(buildS3Key(key), content)
+                .flatMap {
+                    contentSize(key)
+                }
 
-        return s3Client.getObject(request, AsyncResponseTransformer.toPublisher())
-            .toMono()
-            .handleNoSuchKeyException()
-            .flatMapMany { response ->
-                response.toFlux()
-            }
-    }
-
-    override fun upload(key: K, content: Flux<ByteBuffer>): Mono<Long> {
-        val request = CreateMultipartUploadRequest.builder()
-            .bucket(bucketName)
-            .contentType(MediaType.APPLICATION_OCTET_STREAM_VALUE)
-            .key(buildS3Key(key))
-            .build()
-        return s3Client.createMultipartUpload(request)
-            .toMono()
-            .flatMap { response ->
-                content.index()
-                    .flatMap { (index, buffer) ->
-                        response.uploadPart(index + 1, buffer)
-                    }
-                    .collectList()
-                    .flatMap { completedParts ->
-                        val completeRequest = CompleteMultipartUploadRequest.builder()
-                            .bucket(response.bucket())
-                            .key(response.key())
-                            .uploadId(response.uploadId())
-                            .multipartUpload { builder ->
-                                builder.parts(completedParts.sortedBy { it.partNumber() })
-                            }
-                            .build()
-                        s3Client.completeMultipartUpload(completeRequest)
-                            .toMono()
-                    }
-            }
-            .flatMap {
-                contentSize(key)
-            }
-    }
-
-    private fun CreateMultipartUploadResponse.uploadPart(index: Long, contentPart: ByteBuffer): Mono<CompletedPart> {
-        val nextPartRequest = UploadPartRequest.builder()
-            .bucket(bucket())
-            .key(key())
-            .uploadId(uploadId())
-            .partNumber(index.toInt())
-            .build()
-        val nextPartRequestBody = AsyncRequestBody.fromByteBuffer(contentPart)
-        return s3Client.uploadPart(nextPartRequest, nextPartRequestBody)
-            .toMono()
-            .map { partResponse ->
-                CompletedPart.builder()
-                    .eTag(partResponse.eTag())
-                    .partNumber(index.toInt())
-                    .build()
-            }
-    }
-
-    override fun upload(key: K, contentLength: Long, content: Flux<ByteBuffer>): Mono<Unit> {
-        val request = PutObjectRequest.builder()
-            .bucket(bucketName)
-            .contentType(MediaType.APPLICATION_OCTET_STREAM_VALUE)
-            .key(buildS3Key(key))
-            .contentLength(contentLength)
-            .build()
-        return s3Client.putObject(request, AsyncRequestBody.fromPublisher(content))
-            .toMono()
+    override fun upload(key: K, contentLength: Long, content: Flux<ByteBuffer>): Mono<Unit> =
+        s3Operations.uploadObject(buildS3Key(key), contentLength, content)
             .map { response ->
-                log.debug { "Uploaded ${request.bucket()}/${request.key()} with versionId: ${response.versionId()}" }
+                log.debug { "Uploaded $key with versionId: ${response.versionId()}" }
             }
-    }
 
-    override fun delete(key: K): Mono<Boolean> {
-        val request = DeleteObjectRequest.builder()
-            .bucket(bucketName)
-            .key(buildS3Key(key))
-            .build()
-        return s3Client.deleteObject(request)
-            .toMono()
-            .handleNoSuchKeyException()
-            .thenReturn(true)
-            .defaultIfEmpty(false)
-    }
+    override fun delete(key: K): Mono<Boolean> = s3Operations.deleteObject(buildS3Key(key))
+        .thenReturn(true)
+        .defaultIfEmpty(false)
 
-    override fun lastModified(key: K): Mono<Instant> = headObjectAsMono(key)
+    override fun lastModified(key: K): Mono<Instant> = s3Operations.headObject(buildS3Key(key))
         .map { response ->
             response.lastModified()
         }
 
-    override fun contentSize(key: K): Mono<Long> = headObjectAsMono(key)
+    override fun contentSize(key: K): Mono<Long> = s3Operations.headObject(buildS3Key(key))
         .map { response ->
             response.contentLength()
         }
 
-    override fun doesExist(key: K): Mono<Boolean> = headObjectAsMono(key)
+    override fun doesExist(key: K): Mono<Boolean> = s3Operations.headObject(buildS3Key(key))
         .map { true }
         .defaultIfEmpty(false)
-
-    private fun headObjectAsMono(key: K) = HeadObjectRequest.builder()
-        .bucket(bucketName)
-        .key(buildS3Key(key))
-        .build()
-        .let { s3Client.headObject(it) }
-        .toMono()
-        .handleNoSuchKeyException()
 
     /**
      * @param s3KeySuffix cannot start with [PATH_DELIMITER]
@@ -193,10 +93,6 @@ abstract class AbstractS3Storage<K>(
             require(!suffix.startsWith(PATH_DELIMITER)) {
                 "Suffix cannot start with $PATH_DELIMITER: $suffix"
             }
-        }
-
-        private fun <T : Any> Mono<T>.handleNoSuchKeyException(): Mono<T> = onErrorResume(NoSuchKeyException::class.java) {
-            Mono.empty()
         }
     }
 }
